Finished mini Ping program for Phase I of part 2
Need to complete Phase II, and then begin work on ncd again

Now that I ahve a working Ping it should be easy to implement the basic
ncd that I had planned. I should probably look intoplaying with a fork for ping, and maybe even using some signals. They will be useful for the ncd program at the very least.