Removed IPV6 support functions. Begin work on unit tests.
IPV6 support was completely broken. It never worked. I'm stripping it, and will re add this later when NCD works as intended for IPV4.

I've let the unit tests languish for an age. It's time to fix them, even if they aren't that useful. Perhaps I'll refactor the code to be more unit test friendly.

 The biggest problem is, however, that ncd is somewhat ill suited for unit testing. It relies on so many systems, including the internet that true unit testing is sort of pointless.